chore: move api call in separate files
